+++ src/test/kotlin/racingcar/domain/RacingGameTest.kt
@@ -1,9 +1,8 @@
-package step3.domain
+package racingcar.domain
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldBeEmpty
-import racingcar.domain.RacingGame
 import racingcar.model.Car
 import racingcar.model.DefaultCar
